backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.1
supabase==2.10.0
PyJWT==2.9.0
bcrypt==4.2.0
python-multipart==0.0.12
groq==0.11.0
PyPDF2==3.0.1
pydantic==2.9.0
email-validator==2.2.0
reportlab==4.2.2
